requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
einops==0.8.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
MedPy==0.5.2
ml_collections==1.1.0
mpmath==1.3.0
networkx==3.6.1
nibabel==5.4.2
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.2
pillow==12.1.1
protobuf==7.34.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
simpleitk==2.5.3
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.26
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
Werkzeug==3.1.7
